add more spinner behavior
As I use the CLI more, I find the spinner text that gets left behind
isn't actually useful -- it's only useful while it's happening.

After the message was completed, the user will have two states:

It was successful, at which point the informational text that tells
them why they're waiting is no longer useful.
It failed, at which point, we'd also have separate error messages
below the spinner text.
In summary, it's good to have the text in transient form, and clear it
out when done.

== Proposal

1. `Waiting` - used just to indicate that there's activity going on. No
   message necessary since after the wait, the outcome is obvious.

2. `Spinner` - same behavior as before. Useful for scenarios like
   deleting, since there's nothing to show after, and having the
   classic:

   `Deleting application... done`

   Text is better UX.