rules editor create/update tweaks (#229)
* rules editor create/update tweaks

== Description

This adds basic changes for the create/update path. For now we're
keeping them separate with an assertion that we want the straightforward
edit paths to bypass the survey bits -- since that becomes unnecessary
(e.g. when creating a rule, there's only way to do that -- and that's by
editing the rule).

For update, we'll utilize the surveyext bits heavily bollowed from
github to have the `?` prompt.

* Update internal/cli/rules.go

Co-authored-by: Rita Zerrizuela <zeta@widcket.com>

Co-authored-by: Rita Zerrizuela <zeta@widcket.com>