Fix fsnotify and vim issues
TL;DR we need to watch an entire dir instead of just one file.  But also, when
doing that, just utilizing the entire tmpdir is overkill since that's got many
many files in there -- which in practice causes vim to segfault.

The fix is simple: make a tempdir, make the tempfile within that, and clean
those up in reverse order.